feat: extract MDXFromChildren to its own component
